Move the filter buttons to their own component
