+++ shop-app/src/components/app-bar/TopBar.test.tsx
